+++ components/Portfolio/Insurance/DepositHistory/index.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
+import React, { FC } from 'react';
 import { Logo } from '@components/General';
 import { toApproxCurrency } from '@libs/utils';
-import { TableHeading, TableRow, TableCell } from '@components/Portfolio';
+import { TableHeading, TableRow, TableCell } from '@components/General/Table';
 import { DateAndTime } from '@components/General';
 
-const DepositHistory: React.FC = () => {
+const DepositHistory: FC = () => {
     const headings = ['Date', 'Market', 'Amount', 'iTokens Minted', 'Transaction Details'];
 
     const tracers = [
