+++ components/Portfolio/Trading/MarginAccounts/index.tsx
@@ -1,12 +1,13 @@
-import React from 'react';
+import React, { FC } from 'react';
 import { Logo } from '@components/General';
 import { toApproxCurrency } from '@libs/utils';
-import { TableHeading, TableRow, TableCell, Button, StatusIndicator, getStatusColour } from '@components/Portfolio';
+import { TableHeading, TableRow, TableCell } from '@components/General/Table';
+import { Button, StatusIndicator, getStatusColour } from '@components/Portfolio';
 import Tooltip from 'antd/lib/tooltip';
 
 const NoLeverageTip = <p>You have no leveraged trades open in this market.</p>;
 
-const MarginAccounts: React.FC = () => {
+const MarginAccounts: FC = () => {
     const headings = ['Market', 'Equity', 'Maintenance Margin', 'Available Margin', 'Status of Position'];
 
     const tracers = [
