+++ components/Portfolio/Trading/Position/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { FC, useState } from 'react';
 import { Logo } from '@components/General';
 import { toApproxCurrency } from '@libs/utils';
 import {
@@ -6,14 +6,12 @@ import {
     TableLastHeading,
     TableRow,
     TableCell,
-    SecondaryCell,
-    Button,
-    StatusIndicator,
-    getStatusColour,
     TableHeader,
     Table,
-    TableBody, TableLastCell,
-} from '@components/Portfolio';
+    TableBody,
+    TableLastCell,
+} from '@components/General/Table';
+import { SecondaryCell, Button, StatusIndicator, getStatusColour } from '@components/Portfolio';
 import { calcLiquidationPrice, calcUnrealised } from '@tracer-protocol/tracer-utils';
 import { LabelledOrders } from 'libs/types/OrderTypes';
 import { LabelledTracers } from 'libs/types/TracerTypes';
@@ -21,7 +19,7 @@ import { LabelledTracers } from 'libs/types/TracerTypes';
 import TracerLoading from 'public/img/logos/tracer/tracer_loading.svg';
 import Icon from '@ant-design/icons';
 
-const Position: React.FC<{
+const Position: FC<{
     tracers: LabelledTracers;
     allFilledOrders: LabelledOrders;
 }> = ({ tracers, allFilledOrders }) => {
